Create a base git author extractor class
