+++ src/AuthorExtractor/GitAuthorExtractor.php
@@ -20,17 +20,13 @@
 
 namespace ContaoCommunityAlliance\BuildSystem\Tool\AuthorValidation\AuthorExtractor;
 
-use ContaoCommunityAlliance\BuildSystem\Repository\GitException;
 use ContaoCommunityAlliance\BuildSystem\Repository\GitRepository;
-use Symfony\Component\Console\Logger\ConsoleLogger;
-use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Finder\Finder;
-use Symfony\Component\Process\ProcessBuilder;
 
 /**
  * Extract the author information from a git repository.
  */
-class GitAuthorExtractor extends AbstractAuthorExtractor
+class GitAuthorExtractor extends AbstractGitExtractor
 {
     /**
      * Optional attached finder for processing multiple files.
@@ -38,96 +34,6 @@ class GitAuthorExtractor extends AbstractAuthorExtractor
      * @var Finder
      */
     protected $finder;
-
-    /**
-     * Create a git repository instance.
-     *
-     * @param string $path A path within a git repository.
-     *
-     * @return GitRepository.
-     */
-    private function getGitRepositoryFor($path)
-    {
-        $git = new GitRepository($this->determineGitRoot($path));
-        if ($this->output->getVerbosity() >= OutputInterface::VERBOSITY_DEBUG) {
-            $git->getConfig()->setLogger(
-                new ConsoleLogger($this->output)
-            );
-        }
-
-        return $git;
-    }
-
-    /**
-     * Determine the git root, starting from arbitrary directory.
-     *
-     * @param string $path The start path.
-     *
-     * @return string The git root path.
-     *
-     * @throws \RuntimeException If the git root could not determined.
-     */
-    private function determineGitRoot($path)
-    {
-        // @codingStandardsIgnoreStart
-        while (strlen($path) > 1) {
-            // @codingStandardsIgnoreEnd
-            if (is_dir($path . DIRECTORY_SEPARATOR . '.git')) {
-                return $path;
-            }
-
-            $path = dirname($path);
-        }
-
-        throw new \RuntimeException('Could not determine git root, starting from ' . func_get_arg(0));
-    }
-
-    /**
-     * Retrieve a list of all files within a git repository.
-     *
-     * @param GitRepository $git The repository to extract all files from.
-     *
-     * @return string[]
-     *
-     * @throws GitException When the git execution failed.
-     */
-    private function getAllFilesFromGit($git)
-    {
-        $gitDir = $git->getRepositoryPath();
-        // Sadly no command in our git library for this.
-        $processBuilder = new ProcessBuilder();
-        $processBuilder->setWorkingDirectory($gitDir);
-        $processBuilder
-            ->add($git->getConfig()->getGitExecutablePath())
-            ->add('ls-tree')
-            ->add('HEAD')
-            ->add('-r')
-            ->add('--full-name')
-            ->add('--name-only');
-
-        $process = $processBuilder->getProcess();
-
-        $git->getConfig()->getLogger()->debug(
-            sprintf('[ccabs-repository-git] exec [%s] %s', $process->getWorkingDirectory(), $process->getCommandLine())
-        );
-
-        $process->run();
-        $output = rtrim($process->getOutput(), "\r\n");
-
-        if (!$process->isSuccessful()) {
-            throw GitException::createFromProcess('Could not execute git command', $process);
-        }
-
-        $files = array();
-        foreach (explode(PHP_EOL, $output) as $file) {
-            $absolutePath = $gitDir . '/' . $file;
-            if (!$this->config->isPathExcluded($absolutePath)) {
-                $files[trim($absolutePath)] = trim($absolutePath);
-            }
-        }
-
-        return $files;
-    }
 
     /**
      * Convert the git binary output to a valid author list.
@@ -142,21 +48,6 @@ class GitAuthorExtractor extends AbstractAuthorExtractor
             return array();
         }
         return preg_split('~[\r\n]+~', $authors);
-    }
-
-    /**
-     * Retrieve the file path to use in reporting.
-     *
-     * @return string
-     */
-    public function getFilePaths()
-    {
-        $files = array();
-        foreach ($this->config->getIncludedPaths() as $path) {
-            $files = array_merge($files, $this->getAllFilesFromGit($this->getGitRepositoryFor($path)));
-        }
-
-        return $files;
     }
 
     /**
@@ -196,52 +87,6 @@ class GitAuthorExtractor extends AbstractAuthorExtractor
     private function getAuthorListFrom($path, $git)
     {
         return $git->log()->format('%aN <%ae>')->follow()->execute($path);
-    }
-
-    /**
-     * Retrieve the data of the current user on the system.
-     *
-     * @param GitRepository $git The repository to extract all files from.
-     *
-     * @return string
-     *
-     * @throws GitException When the git execution failed.
-     */
-    private function getCurrentUserInfo($git)
-    {
-        // Sadly no command in our git library for this.
-        $processBuilder = new ProcessBuilder();
-        $processBuilder->setWorkingDirectory($git->getRepositoryPath());
-        $processBuilder
-            ->add($git->getConfig()->getGitExecutablePath())
-            ->add('config')
-            ->add('--get-regexp')
-            ->add('user.[name|email]');
-
-        $process = $processBuilder->getProcess();
-
-        $git->getConfig()->getLogger()->debug(
-            sprintf('[ccabs-repository-git] exec [%s] %s', $process->getWorkingDirectory(), $process->getCommandLine())
-        );
-
-        $process->run();
-        $output = rtrim($process->getOutput(), "\r\n");
-
-        if (!$process->isSuccessful()) {
-            throw GitException::createFromProcess('Could not execute git command', $process);
-        }
-
-        $config = array();
-        foreach (explode(PHP_EOL, $output) as $line) {
-            list($name, $value)  = explode(' ', $line, 2);
-            $config[trim($name)] = trim($value);
-        }
-
-        if (isset($config['user.name']) && $config['user.email']) {
-            return sprintf('%s <%s>', $config['user.name'], $config['user.email']);
-        }
-
-        return '';
     }
 
     /**
